Run As Root Checks

+++ uninstall.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-
+if [[ $(id -u) -ne 0 ]] ; then echo "Please run as root" ; exit 1 ; fi
 gpasswd -d peapod music
 userdel peapod;
 rm -fr /home/peapod;
